+++ demo-app/src/main/kotlin/io/getstream/video/android/util/filters/BlurredBackgroundVideoFilter.kt
@@ -1,0 +1,76 @@
+/*
+ * Copyright (c) 2014-2024 Stream.io Inc. All rights reserved.
+ *
+ * Licensed under the Stream License;
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    https://github.com/GetStream/stream-video-android/blob/main/LICENSE
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.getstream.video.android.util.filters
+
+import android.graphics.Bitmap
+import android.graphics.Canvas
+import com.google.android.gms.tasks.Tasks
+import com.google.android.renderscript.Toolkit
+import com.google.mlkit.vision.common.InputImage
+import com.google.mlkit.vision.segmentation.Segmentation
+import com.google.mlkit.vision.segmentation.SegmentationMask
+import com.google.mlkit.vision.segmentation.selfie.SelfieSegmenterOptions
+import io.getstream.video.android.core.call.video.BitmapVideoFilter
+
+/**
+ * Applies a blur effect to the background of a video frame.
+ *
+ * Note that this filter is still in beta and may not work as expected. To tweak it, see constants at bottom of file.
+ */
+class BlurredBackgroundVideoFilter : BitmapVideoFilter() {
+    private val options =
+        SelfieSegmenterOptions.Builder()
+            .setDetectorMode(SelfieSegmenterOptions.STREAM_MODE)
+            .enableRawSizeMask()
+            .build()
+    private val segmenter = Segmentation.getClient(options)
+
+    private lateinit var segmentationMask: SegmentationMask
+    private val onlyBackgroundBitmap by lazy {
+        Bitmap.createBitmap(
+            segmentationMask.width,
+            segmentationMask.height,
+            Bitmap.Config.ARGB_8888,
+        )
+    }
+
+    override fun filter(videoFrameBitmap: Bitmap) {
+        // 1. Apply segmentation
+        // 2. Copy the background segment to a new bitmap
+        // 3. Blur the background bitmap
+        // 4. Draw the blurred background bitmap on the original bitmap
+
+        val mlImage = InputImage.fromBitmap(videoFrameBitmap, 0)
+        val task = segmenter.process(mlImage)
+        segmentationMask = Tasks.await(task)
+
+        copySegment(
+            segment = Segment.BACKGROUND,
+            source = videoFrameBitmap,
+            destination = onlyBackgroundBitmap,
+            segmentationMask = segmentationMask,
+        )
+
+        val blurredBackgroundBitmap = Toolkit.blur(onlyBackgroundBitmap, BLUR_RADIUS.toInt())
+        val canvas = Canvas(videoFrameBitmap)
+        val matrix = newSegmentationMaskMatrix(videoFrameBitmap, segmentationMask)
+
+        canvas.drawBitmap(blurredBackgroundBitmap, matrix, null)
+    }
+}
+
+private const val BLUR_RADIUS = 10f // Set the radius of the Blur. Supported range 0 < radius <= 25
